Add Operators to UI Menu
The only other place to find the operators is under a dropdown menu, then clicking the slider button, or by fimbling through the number hotkeys til you find an operator that appears to be what you wanted (or the pie menus, which are also only under a hotkey).

They are sorted mostly by how they are in the keymap. They could be ordered more by functionality and/or name.