Remove hotkeys and their description in operators
In addition to Blender 2.8 guideline for addons to not cram hotkeys onto the user (and just allow them to add the hotkeys they want), there are too many hotkeys to remember and finding the related operator is a matter of fumbling around until you find it.

Currently, the problem with this method after removing default keys, is the user cannot add hotkeys to anything other than the pie menus.

I can only guess the reasoning behind this, so I've no alternatives and only one solution.
When an operator's first world in its bl_idname, is a custom one ('animaide.tween' ) instead of one of the premade internal ones ('graph.tween', 'anim.tween'),
the operator can run and be added as a hotkey via python/manual insertion,
but when displayed in the UI, the right-click menu does not offer an Assign Shortcut button.