Fix error when running Blend Neighbor (with presets)
When run from invoke, the operator inserts a global variable that's used then used in the slider function.
Running from execute, skips creating these.

I mistook the data from the prior keymaps, as the same thing as operator context, when it's actually just a string/bool variable to send from invoke, to tell a function whether or not to run in its entirety